requirements.txt
absl-py==0.9.0
biopython==1.76
bleach==1.5.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
ete3==3.1.1
h5py==2.10.0
html5lib==0.9999999
idna==2.8
joblib==0.14.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
matplotlib==3.1.3
networkx==2.4
numpy==1.18.0
pandas==0.25.3
protobuf==3.11.2
pygraphviz==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
spektral==0.1.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
urllib3==1.25.7
Werkzeug==0.16.0
